% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare_wvar}
\alias{compare_wvar}
\title{Comparison between classical and robust Wavelet Variances}
\usage{
compare_wvar(..., nb_ticks_x = NULL, nb_ticks_y = NULL)
}
\arguments{
\item{...}{One or more time series objects.}

\item{nb_ticks_x}{Maximum number of ticks for the x-axis.}

\item{nb_ticks_y}{Maximum number of ticks for the y-axis.}
}
\description{
Displays a plot of the wavelet variances (classical and robust) for a given time series accounting for CI values.
}
