% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{der_psi_tuk}
\alias{der_psi_tuk}
\title{Derivative of Psi Tukey Scoring function}
\usage{
der_psi_tuk(x, sig2_bw, crob_bw)
}
\arguments{
\item{x}{data vector}

\item{sig2_bw}{variance of brickwalled wv coefficients}

\item{crob_bw}{?}
}
\description{
Derivative of Psi Tukey Scoring function
}
