% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.R
\name{dwt}
\alias{dwt}
\title{Discrete Wavelet Transform}
\usage{
dwt(x, nlevels = floor(log2(length(x))), filter = "haar",
  boundary = "periodic", bw = TRUE)
}
\value{
y A \code{field<vec>} that contains the wavelet coefficients for each decomposition level
}
\description{
Calculation of the coefficients for the discrete wavelet transformation
}
\details{
Performs a level \eqn{J} decomposition of the time series using the pyramid algorithm.
The default \eqn{J} is determined by \eqn{floor\left(log_2 \left(length\left(x\right)\right)\right)}{floor(log2(length(x)))}
}
\examples{
set.seed(999)
x = rnorm(2^8)
dwt(x)
}
\author{
JJB
}
