% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dwt_cpp}
\alias{dwt_cpp}
\title{Discrete Wavelet Transform}
\usage{
dwt_cpp(x, filter_name, nlevels)
}
\arguments{
\item{x}{A \code{vector} with dimensions \eqn{N\times 1}{N x 1}.}

\item{filter_name}{A \code{string} indicating the filter.}

\item{nlevels}{An \code{integer}, \eqn{J}, indicating the level of the decomposition.}
}
\value{
y A \code{field<vec>} that contains the wavelet coefficients for each decomposition level
}
\description{
Calculation of the coefficients for the discrete wavelet transformation.
}
\details{
Performs a level J decomposition of the time series using the pyramid algorithm
}
\author{
James Balamuta and Justin Lee
}
\keyword{internal}
