% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\name{ggplot_like_colors}
\alias{ggplot_like_colors}
\title{Get ggplot2-like colors}
\usage{
ggplot_like_colors(n, alpha = 1)
}
\arguments{
\item{n}{number of colors.}

\item{alpha}{transparency.}
}
\value{
list of colors.
}
\description{
Get ggplot2-like colors
}
