% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hfilter}
\alias{hfilter}
\title{Haar filter for a spatial case}
\usage{
hfilter(jscale)
}
\arguments{
\item{jscale}{An \code{int} of the Number of Scales}
}
\description{
Haar filter for a spatial case
}
