% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_mat}
\alias{index_mat}
\title{Create an index matrix}
\usage{
index_mat(m, n, ftype)
}
\arguments{
\item{m, n}{An \code{integer} indicating matrix row, column dimension}

\item{ftype}{A \code{string} of either \code{"col"} or \code{"row"}.}
}
\description{
Create an index matrix
}
\details{
Creates an index matrix filled by either row or column.
}
\examples{
# Index matrix by column
index_mat(3,4, "col")
# index matrix by row
index_mat(3,4, "row")
}
