% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lower_tri_elem}
\alias{lower_tri_elem}
\title{Extract Lower Triangular Elements}
\usage{
lower_tri_elem(X)
}
\arguments{
\item{X}{\code{matrix}}
}
\description{
Extract Lower Triangular Elements
}
\examples{
x = matrix(1:16, 4, 4, byrow = TRUE)
lower_tri_elem(x)
}
