% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_wv_iso}
\alias{make_wv_iso}
\title{Create the ISO matrix}
\usage{
make_wv_iso(wv, min_dim)
}
\arguments{
\item{wv}{A \code{matrix} containing the wavelet variance.}

\item{min_dim}{An \code{integer} indicating the minimum value.}
}
\description{
Create the ISO matrix
}
\details{
Takes the average of the lower triangular view
before returning.
}
\examples{
a = matrix(1:9, nrow = 3, byrow = TRUE)
make_wv_iso(a, 3 - 1)
}
