% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modwt.R
\name{modwt}
\alias{modwt}
\title{Maximum Overlap Discrete Wavelet Transform}
\usage{
modwt(x, nlevels = floor(log2(length(x) - 1)), filter = "haar")
}
\arguments{
\item{x}{A \code{vector} with dimensions N x 1.}

\item{nlevels}{A \code{integer} indicating the \eqn{J} levels of decomposition.}

\item{filter}{A \code{string} indicating the filter name}
}
\value{
A \code{field<vec>} that contains the wavelet coefficients for each decomposition level
}
\description{
Calculates the coefficients for the discrete wavelet transformation
}
\details{
Performs a level \eqn{J} decomposition of the time series using the pyramid algorithm.
The default \eqn{J} is determined by \eqn{floor\left(log_2 \left(length\left(x\right)\right)\right)}{floor(log2(length(x)))}
}
\examples{
set.seed(999)
x = rnorm(100)
ret = modwt(x)

summary(ret)

plot(ret)
}
\author{
James Balamuta, Justin Lee and Stephane Guerrier
}
