% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{modwt_cpp}
\alias{modwt_cpp}
\title{Maximum Overlap Discrete Wavelet Transform}
\usage{
modwt_cpp(x, filter_name, nlevels)
}
\arguments{
\item{x}{A \code{vector} with dimensions \eqn{N\times 1}{N x 1}.}

\item{filter_name}{A \code{string} indicating the filter.}

\item{nlevels}{An \code{integer}, \eqn{J}, indicating the level of the decomposition.}
}
\value{
y A \code{field<vec>} that contains the wavelet coefficients for each decomposition level
}
\description{
Calculation of the coefficients for the discrete wavelet transformation
}
\details{
Performs a level J decomposition of the time series using the pyramid algorithm.
Use this implementation to supply custom parameters instead of modwt(x),
which serves as a wrapper function.
}
\author{
James Balamuta and Justin Lee
}
\keyword{internal}
