% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo_noav.R
\name{NOAV}
\alias{NOAV}
\title{Non-stationary Non-overlapping Allan Variance}
\usage{
NOAV(n, covmat)
}
\arguments{
\item{n}{A \code{integer} indicating the length of each vector of consecutive observations considered for the average.}

\item{covmat}{A \code{matrix} indicating the T-by-T covariance matrix of the time series with length T.}
}
\value{
A \code{field<numeric>} that is the theoritical Non-overlapping Allan Variance for constant-mean non-stationary time series data
}
\description{
Calculation of the theoritical Non-overlapping Allan Variance for constant-mean non-stationary time series data
}
\details{
Calculate Non-overlapping Allan Variance based on the defination on "Xu, Haotian, et al. "A Study of the Allan Variance for Constant-Mean Non-Stationary Processes." arXiv preprint arXiv:1702.07795 (2017)".
n is an integer larger than 1 and smaller than \eqn{floor\left(log_2 \left(dim\left(covmat\right)[1]\right)\right)-1}{floor(log2(dim(T)[1]))-1}
}
\examples{
set.seed(999)
AR1 phi=0.3
x = gen.gts(AR1(phi = 0.3, sigma2 = 1), N = 1000, freq = 1)
avar(x, type = "to")

a = matrix(rep(0, 1000^2), nrow = 1000)
for (i in 1:1000){
  a[,i] = seq(from = 1 - i, length.out = 1000)
}
a.diag = diag(a)
a[upper.tri(a,diag=T)] = 0
a = a + t(a) + diag(a.diag)
covmat = 0.3^a
sapply(1:8, function(y){NOAV(2^y, covmat)})
}
\author{
Haotian Xu
}
