% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fun.R
\name{plot.dwt}
\alias{plot.dwt}
\title{Plot function for dwt objects.}
\usage{
\method{plot}{dwt}(object, index = NULL, couleur = NULL)
}
\arguments{
\item{\code{y}}{A \code{dwt} object.}

\item{\code{index}}{A \code{vector} containing the indices to scales to be included in 
the graph. By default \code{index = 1:(min(c(J,4)))}, where \code{J} denotes the 
number of scales in \code{y}.}

\item{\code{couleur}}{A \code{vector} of colors of the same size as \code{index} used
for the different scales depicted in the graph. If \code{couleur} contains a single 
value thn the same color will be used for all scales.}
}
\description{
Plot function for dwt objects.
}
\examples{
# Simulate a Gaussian white noise
n = 10^3
Xt = rnorm(n)

# dwt
Xt.dwt = dwt(Xt)

# Graph examples
plot(Xt.dwt)
plot(Xt.dwt, index = c(1,4,5,6,8,2))
plot(Xt.dwt, index = c(1,4,5,6), couleur = "blue")
plot(Xt.dwt, index = c(1,4,5,6), couleur = rep(c("blue","yellow"),2))
}
\author{
Justin Lee and Stéphane Guerrier
}
