% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covwv.R
\name{plot.wccv}
\alias{plot.wccv}
\title{Plot Cross Covariance}
\usage{
\method{plot}{wccv}(x, theo.wccv = NULL)
}
\description{
Plots results of the wccv list in which additional parameters can be specified
}
\author{
Haotian Xu, Stephane Guerrier, and Justin Lee
}
\keyword{internal}
