% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covwv.R
\name{plot.wccv_pair}
\alias{plot.wccv_pair}
\title{Plot Cross Covariance Pair}
\usage{
\method{plot}{wccv_pair}(x, theo.wccv = NULL, ...)
}
\description{
Plots results of the a wccv_pair list in which additional parameters can be specified
}
\author{
Haotian Xu, Stephane Guerrier, and Justin Lee
}
