% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{plot.wvar}
\alias{plot.wvar}
\title{Plot Wavelet Variances}
\usage{
\method{plot}{wvar}(x, units = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col_wv = NULL, col_ci = NULL, nb_ticks_x = NULL,
  nb_ticks_y = NULL, legend_position = NULL, ...)
}
\arguments{
\item{\code{x}}{A \code{wvar} object.}

\item{\code{xlab}}{A title for the x axis.}

\item{\code{ylab}}{A title for the y axis.}

\item{\code{main}}{An overall title for the plot.}

\item{\code{col_wv}}{Color of the wavelet variance line.}

\item{\code{col_ci}}{Color of the confidence interval shade.}

\item{\code{nb_ticks_x}}{Number of ticks for the x-axis.}

\item{\code{nb_ticks_y}}{Number of ticks for the y-axis.}

\item{\code{legend_position}}{Position of the legend.}

\item{\code{...}}{Additional arguments affecting the plot.}
}
\value{
Plot of wavelet variance and confidence interval for each scale.
}
\description{
Displays a plot of wavelet variance accounting for CI values and supplied efficiency.
}
\author{
Stephane Gurrier, Nathanael Claussen, and Justin Lee
}
\keyword{internal}
