% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_tuk}
\alias{psi_tuk}
\title{Psi Tukey Scoring function}
\usage{
psi_tuk(x, sig2_bw, crob_bw)
}
\arguments{
\item{x}{data vector}

\item{sig2_bw}{variance of brickwalled wv coefficients}

\item{crob_bw}{?}
}
\value{
A \code{vec} containing the scores.
}
\description{
Psi Tukey Scoring function
}
