% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_cpp}
\alias{quantile_cpp}
\title{Find Quantiles}
\usage{
quantile_cpp(x, probs)
}
\arguments{
\item{x}{A \code{vec} of data}

\item{probs}{A \code{vec} of the quantiles to find.}
}
\value{
A \code{vector} containing the quantiles
}
\description{
Attempts to find quantiles
}
\author{
James J Balamuta
}
\keyword{internal}
