% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{robust_eda}
\alias{robust_eda}
\title{Comparison between classical and robust Wavelet Variances}
\usage{
robust_eda(x, eff = 0.6, units = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col_wv = NULL, col_ci = NULL, nb_ticks_x = NULL,
  nb_ticks_y = NULL, legend_position = NULL, ...)
}
\arguments{
\item{eff}{The efficiency of the robust estimator}

\item{{x}}{A time series object.}

\item{{xlab}}{A title for the x axis.}

\item{{ylab}}{A title for the y axis.}

\item{{main}}{An overall title for the plot.}

\item{{col_wv}}{Color of the wavelet variance line.}

\item{{col_ci}}{Color of the confidence interval shade.}

\item{{nb_ticks_x}}{Maximum number of ticks for the x-axis.}

\item{{nb_ticks_y}}{Maximum number of ticks for the y-axis.}

\item{{legend_position}}{Position of the legend (use legend_position = NA to remove legend).}

\item{{...}}{Additional arguments affecting the plot.}
}
\value{
Plot of wavelet variance and confidence interval for each scale.
}
\description{
Displays a plot of the wavelet variances (classical and robust) for a given time series accounting for CI values.
}
\examples{
n = 10^4
Xt = rnorm(n)
wv = wvar(Xt)
plot(wv)
plot(wv, main = "Simulated white noise", xlab = "Scales")
plot(wv, units = "sec", legend_position = "topright")
plot(wv, col_wv = "darkred", col_ci = "pink")
}
\author{
Stephane Guerrier, Nathanael Claussen, and Justin Lee
}
