% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_modwt_cpp}
\alias{sp_modwt_cpp}
\title{Compute the Spatial Wavelet Coefficients}
\usage{
sp_modwt_cpp(X, J1, J2)
}
\arguments{
\item{X}{is a matrix with row, col orientation}

\item{J1, J2}{is the levels of decomposition along the rows, columns}
}
\value{
A \code{list} of \code{vectors} containing the wavelet coefficients.
}
\description{
Compute the Spatial Wavelet Coefficients
}
\details{
By default this function will return the wavelet coefficient in
addition to the wavelet
}
