% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_wvar_cpp}
\alias{sp_wvar_cpp}
\title{Spatial Wavelet Variance}
\usage{
sp_wvar_cpp(wv_coeffs, n, m, J1, J2, iso = TRUE, robust = TRUE, eff = 0.6)
}
\arguments{
\item{wv_coeffs}{is \code{field<vec>} containing the wavelet coefficients}

\item{n, m}{is \code{int} containing the number of observations in rows, cols}

\item{J1}{is the levels of decomposition along the rows, columns}

\item{J2}{is the levels of decomposition along the rows, columns}

\item{iso}{is whether the matrix is isometric}

\item{robust}{\code{bool} is an indicator as to whether a classic or robust
estimation should occur.}

\item{eff}{\code{double} is the level of efficiency}
}
\description{
Computes the Spatial Wavelet Variance
}
