% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spat_wavar}
\alias{spat_wavar}
\title{Compute the Spatial Wavelet Variance}
\usage{
spat_wavar(X, J1, J2, iso = TRUE, robust = TRUE, eff = 0.6)
}
\arguments{
\item{X}{is a matrix with row, col orientation}

\item{J1}{is the levels of decomposition along the rows, columns}

\item{J2}{is the levels of decomposition along the rows, columns}

\item{iso}{is whether the matrix is isometric}

\item{robust}{\code{bool} is an indicator as to whether a classic or robust
estimation should occur.}

\item{eff}{\code{double} is the level of efficiency}
}
\value{
A \code{list} with the following values:
\itemize{
\item{Classical WV}
\item{Robust WV}
\item{Covariance Matrix}
}
}
\description{
Compute the Spatial Wavelet Variance
}
\examples{
eps = 0 # no contamination
eps = 0.05 # with contamination
sig.eps = 9
m = n = 10
nlag = seq(1,n,1)
grid = expand.grid(nlag,nlag)
J1 = floor(log2(n)) - 1
J2 = floor(log2(m)) - 1

# Parameters
sig2 = 1
phi = 2
theta = c(phi,sig2)
wv.theo = exp_theo(theta,J1,J2)
p = length(theta)
alpha = 0.05
J = J1*(J2+1)/2

# Covariance matrix for simulation of process
distM = as.matrix(dist(grid))
Sigma = sig2*exp(-distM/phi)
cholSigma = t(chol(Sigma))

# Simulate the matrix
set.seed(234)
sim = cholSigma\%*\%rnorm(m*n)

# If contamination is set, contaminate
if(eps!=0){
    index = sample(1:(m*n),round((m*n)*eps))
    sim[index] = sim[index] + rnorm(length(index),0,sqrt(sig.eps))
}

Ymle = sim

Ygmwm = matrix(sim,n)

wv = spat_wavar(Ygmwm, J1, J2, eff = 0.6)
}
\seealso{
\code{\link{sp_wv_coeffs}}, \code{\link{spgmwm_exp}}
}
