% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_matrix_parallel}
\alias{subset_matrix_parallel}
\title{Perform parallel contingous subset}
\usage{
subset_matrix_parallel(x, y)
}
\arguments{
\item{x}{\code{matrix} to use to subset}

\item{y}{\code{vector} of ids}
}
\description{
Perform parallel contingous subset
}
\examples{
a = matrix(1:100, nrow = 10, byrow = TRUE)

idx = c(1, 3, 10)
subset_matrix_parallel(a, idx - 1)
## same as
# a[c(1, 3, 10), c(1, 3, 10)]
}
