% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modwt.R
\name{summary.modwt}
\alias{summary.modwt}
\title{Summary Maximum Overlap Discrete Wavelet Transform}
\usage{
\method{summary}{modwt}(object, ...)
}
\arguments{
\item{object}{A \code{modwt} object}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Prints the modwt matrix decomposition
}
\description{
Prints MODWT object in a concise format
}
\examples{
set.seed(999)
x = rnorm(100)
summary(modwt(x))
}
\author{
Nathanael Claussen and Justin Lee
}
\keyword{internal}
