% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{summary.wvar}
\alias{summary.wvar}
\title{Summary of Wavelet Variances}
\usage{
\method{summary}{wvar}(object, ...)
}
\arguments{
\item{object}{A \code{wvar} object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
Summary table and other properties of the object.
}
\description{
Displays the summary table of wavelet variance accounting for CI values and supplied efficiency.
}
\examples{
set.seed(999)
x = rnorm(100)
ret = wvar(x)
summary(ret)
}
\author{
James Balamuta
}
\keyword{internal}
