% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare_wvar_split}
\alias{compare_wvar_split}
\title{Comparison between multiple Wavelet Variances}
\usage{
compare_wvar_split(..., nb_ticks_x = NULL, nb_ticks_y = NULL)
}
\arguments{
\item{...}{One or more time series objects.}

\item{nb_ticks_x}{An \code{integer} that specifies the maximum number of ticks for the x-axis.}

\item{nb_ticks_y}{An \code{integer} that specifies the maximum number of ticks for the y-axis.}
}
\description{
Displays plots of multiple wavelet variances of different time series accounting for CI values.
}
\examples{
set.seed(999)
n = 10^4
Xt = arima.sim(n = n, list(ar = 0.10))
Yt = arima.sim(n = n, list(ar = 0.35))
Zt = arima.sim(n = n, list(ar = 0.70))
Wt = arima.sim(n = n, list(ar = 0.95))

wv_Xt = wvar(Xt)
wv_Yt = wvar(Yt)
wv_Zt = wvar(Zt)
wv_Wt = wvar(Wt)

compare_wvar(wv_Xt, wv_Yt, wv_Zt, wv_Wt)
}
\author{
Stephane Guerrier and Justin Lee
}
