% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{compare_wvar_split}
\alias{compare_wvar_split}
\title{Multi-Plot Comparison Between Multiple Wavelet Variances}
\usage{
compare_wvar_split(graph_details)
}
\arguments{
\item{graph_details}{List of inputs}
}
\description{
This is a helper function for the \code{compare_var()} function. 
This method accepts the same set of arguments as \code{compare_wvar} and returns a comparision 
of multiple wavelet variances of different time series accounting for CI values as a set of different plots.
}
\author{
Stephane Guerrier, Justin Lee, and Nathanael Claussen
}
