% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{plot.wvar}
\alias{plot.wvar}
\title{Plot Wavelet Variances}
\usage{
\method{plot}{wvar}(x, ...)
}
\arguments{
\item{x}{A \code{wvar} object.}

\item{...}{additional arguments affecting the plot produced.}
}
\value{
plot of the object.
}
\description{
Displays plot of wavelet variance accounting for CI values and supplied efficiency.
}
\examples{
set.seed(999)
x = rnorm(1000)
ret = wvar(x)
summary(ret)
}
\author{
James Balamuta
}
\keyword{internal}
