% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{plot.wvar}
\alias{plot.wvar}
\title{Plot Wavelet Variances}
\usage{
\method{plot}{wvar}(x, units = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col_wv = NULL, col_ci = NULL, nb_ticks_x = NULL,
  nb_ticks_y = NULL, legend_position = NULL, ...)
}
\arguments{
\item{x}{A \code{wvar} object.}

\item{col_wv}{Color of wavelet variance line.}

\item{col_ci}{Color of Confidence Interval shade.}

\item{nb_ticks_x}{Number of ticks for the x-axis.}

\item{nb_ticks_y}{Number of ticks for the y-axis.}

\item{legend_position}{Position of the Legend.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Displays plot of wavelet variance accounting for CI values and supplied efficiency.
}
\keyword{internal}
