% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvar.R
\name{plot.wvar}
\alias{plot.wvar}
\title{Plot Wavelet Variances}
\usage{
\method{plot}{wvar}(x, units = NULL, xlab = NULL, ylab = NULL,
  main = NULL, col_wv = NULL, col_ci = NULL, nb_ticks_x = NULL,
  nb_ticks_y = NULL, legend_position = NULL, ...)
}
\arguments{
\item{{x}}{A \code{wvar} object.}

\item{{xlab}}{A title for the x axis.}

\item{{ylab}}{A title for the y axis.}

\item{{main}}{An overall title for the plot.}

\item{{col_wv}}{Color of the wavelet variance line.}

\item{{col_ci}}{Color of the confidence interval shade.}

\item{{nb_ticks_x}}{Number of ticks for the x-axis.}

\item{{nb_ticks_y}}{Number of ticks for the y-axis.}

\item{{legend_position}}{Position of the legend.}

\item{{...}}{Additional arguments affecting the plot.}
}
\value{
Plot of wavelet variance and confidence interval for each scale.
}
\description{
Displays a plot of wavelet variance accounting for CI values and supplied efficiency.
}
\author{
Stephane Guerrier, Nathanael Claussen, and Justin Lee
}
\keyword{internal}
