% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_modwt.R
\name{sp_modwt}
\alias{sp_modwt}
\title{Maximum Overlap Discrete Wavelet Transform}
\usage{
sp_modwt(x, J1 = NULL, J2 = NULL)
}
\arguments{
\item{x}{A \code{matrix} with dimensions N x M.}

\item{J1}{A \code{integer} indicating the \eqn{J1} row levels of decomposition.}

\item{J2}{A \code{integer} indicating the \eqn{J2} column levels of decomposition}
}
\value{
A \code{field<vec>} that contains the wavelet coefficients for each decomposition level
}
\description{
Calculates the coefficients for the discrete wavelet transformation
}
\author{
Justin Lee
}
