% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_modwt}
\alias{sp_modwt}
\title{Compute the Spatial Wavelet Coefficients}
\usage{
sp_modwt(X, J1 = floor(log2(dim(X)[1] - 1)), J2 = floor(log2(dim(X)[2] -
  1)))
}
\arguments{
\item{X}{is a matrix with row, col orientation}

\item{J1, J2}{is the levels of decomposition along the rows, columns}
}
\value{
A \code{list} of \code{vectors} containing the wavelet coefficients.
}
\description{
Compute the Spatial Wavelet Coefficients
}
\details{
By default this function will return the wavelet coefficient in
addition to the wavelet
}
